test(skills): keep skill_entry unit tests with tools/, move the shell journey next to the others
tests/tools/interactive_shell/ is the existing home for tools/interactive_shell
modules (test_session_goal_action.py, test_action_names.py). The one test
that drives the surface's run_action_tool_turn now sits with the other
onboarding journeys in tests/interactive_shell/runtime/test_demo_picker.py.
